ajout de la fonction de récupération de couleur à partir d'une configuration sur les fonds Ajout du logo du jeux dans la page de chargement Ajout du helper de fichier (nombre d'élements dans le dossier , élements aléatoires du dossier) Début d'implémentation de l'animation d'affichage personnages (page de chargement) Ajout des dossiers personnage de test Récupération des données personnages à faire et animation page de loading à faire

+++ src/game/scene/background/game-background/GridBackground.hpp
@@ -32,7 +32,7 @@ namespace Game::Scene::Background::GameBackground{
              * @param newSquareSize taille des carrés
              * @return self
              */
-            GridBackground* setSquareSize(int newSquareSize);
+            GridBackground* setSquareSize(float newSquareSize);
 
             SceneBackground* drawIn(Rectangle container) noexcept override;
 
@@ -46,7 +46,7 @@ namespace Game::Scene::Background::GameBackground{
              * @param drawSquareSize taille du carré
              * @return self
              */
-            GridBackground* drawSquareIn(Rectangle container,Color color,int drawSquareSize);
+            GridBackground* drawSquareIn(Rectangle container,Color color,float drawSquareSize);
 
         private:
             /**
@@ -57,7 +57,7 @@ namespace Game::Scene::Background::GameBackground{
             /**
              * taille des carrées de la grille
              */
-            int squareSize;
+            float squareSize;
     };
 }
